fixed index load simulator button

+++ app/View/ProgramParticipants/index.ctp
@@ -182,7 +182,7 @@
 	                           __('Are you sure you want to delete participant %s ?', $participant['Participant']['phone'])); ?>
 	                   <?php } ?>
 	                   <?php
-	                   if (isset($participant['Participant']['simulate'])) {
+	                   if (isset($participant['Participant']['simulate']) && ($participant['Participant']['simulate']) ) {
 	                       echo $this->Html->link(__('Simulate'), array('program' => $programDetails['url'], 'controller' => 'programParticipants', 'action' => 'simulateMo', $participant['Participant']['_id']));
 	                   }
 	                   ?>
